+++ MathematicaInformationTools/PacletInfo.m
@@ -1,0 +1,16 @@
+(* Mathematica Source File *)
+(* Created by Mathematica Plugin for IntelliJ IDEA *)
+(* :Author: patrick *)
+(* :Date: 2018-09-30 *)
+
+Paclet[
+  Name -> "MathematicaInformationTools",
+  Version -> "0.1",
+  MathematicaVersion -> "11+",
+  Description -> "",
+  Creator -> "Patrick Scheibe",
+  Extensions ->
+      {
+        {"Kernel", Root -> "Kernel", Context -> "MathematicaInformationTools`"}
+      }
+]