+++ resources/mathematica/FunctionInformation.m
@@ -20,20 +20,33 @@ CreateSymbolVersions::usage = "CreateSymbolVersions[] creates a list all symbols
 sym2 -> $VersionNumber}. CreateSymbolVersions[list] updates a list of versioned symbol with the current $VersionNumber.";
 InitializeSymbolInformation::usage = "InitializeSymbolInformation[] loads all symbols we need in the Plugin for completion and resolving";
 
+$allNames;
+isFunction;
+getOptions;
+getAttributes;
+$functionFrequency;
+$functionInformation;
+$symbolVersions;
+
 Begin["`Private`"];
-
-$additionalSymbols = {
-  "FEPrivate`AddSpecialArgCompletion"
-}
-
 << JLink`;
+
+$symbolVersions = Get[FileNameJoin[{DirectoryName@System`Private`$InputFileName, "versionedSymbols.m"}]];
+
+$contexts = Sort[Join[Select[Contexts[], StringFreeQ[#, {
+  "`Private`",
+  "`PackagePrivate`",
+  StartOfString ~~ _?LowerCaseQ ~~ ___,
+  StartOfString ~~ "$" ~~ ___
+}] &], {"System`Private`"}]];
+
 
 (* For good code completion we need an ordering of all possible completions. This is done with the *)
 (* function frequency list that comes with Mathematica nowadays. I just assign numbers according to the *)
 (* place in this list. The higher the number, the more important and the more like is the completion result. *)
 $functionFrequency = With[{file = First[FileNames["all_top_level.m", {$InstallationDirectory}, Infinity]]},
   Dispatch[Append[
-    MapIndexed[Rule[#1, ToString @@ #2]&, Reverse[Get[file]]],
+    MapIndexed[Rule["System`" <> #1, First[#2]]&, Reverse[Get[file]]],
     _ -> "0"
   ]]
 ];
@@ -48,17 +61,15 @@ makeContextNames[context_String] := Block[{$ContextPath = {context}},
 ];
 
 InitializeSymbolInformation[] := Module[{},
-  $builtInNames = Sort[Flatten[ makeContextNames /@ {"System`"} ]];
-  $versionedNames = Sort[Flatten[ makeContextNames /@ {"System`", "Developer`", "Internal`", "JLink`"} ]];
-  $allNames = Sort[Flatten[ makeContextNames /@ DeleteCases[Contexts[], "Global`" | "System`"]]];
+  $allNames = Sort[Flatten[ makeContextNames /@ $contexts]];
 ]
 
 CreateAuxNames[outDir_ /; DirectoryQ[outDir]] := Module[
   {
     contexts = Union[Context /@ $allNames]
   },
-  Export[FileNameJoin[{outDir, "contexts.properties"}], contexts, "Table"];
-  Export[FileNameJoin[{outDir, "contextSymbols.properties"}], Sort[Join[$additionalSymbols, $allNames]], "Table"];
+  Export[FileNameJoin[{outDir, "MathematicaContexts.properties"}], contexts, "Table"];
+  Export[FileNameJoin[{outDir, "MathematicaContextSymbols.properties"}], Sort[Join[$additionalSymbols, $allNames]], "Table"];
 ];
 
 CreateSymbolVersions[] := Thread[$versionedNames -> $VersionNumber];
@@ -66,51 +77,12 @@ CreateSymbolVersions[existingNames_List] := With[{currVersion = $VersionNumber},
   existingNames /. (Function[n, (n -> oldVersion_) :> (n -> Min[{currVersion, oldVersion}])] /@ $versionedNames)
 ];
 
-isFunction[str_String] :=
-    With[{usg =
-        ToString[
-          Function[s, MessageName[s, "usage"], HoldAll] @@
-              ToHeldExpression[str]]},
-      str <> " " <>
-          If[StringMatchQ[usg, __ ~~ str ~~ "[" ~~ ___ ~~ "]" ~~ ___],
-            " = true", " = false"]
-    ]
-
-getOptions[str_String] :=
-    str <> " = " <>
-        StringTrim[
-          Function[expr,
-            Riffle[ToString[#, InputForm] & /@ (First /@
-                Options[Unevaluated[expr]]), " "] // StringJoin, HoldAll] @@
-              ToHeldExpression[str], "{" | "}" | ","]
-
-getAttributes[str_String] :=
-    str <> " = " <>
-        StringTrim[
-          Function[expr,
-            Riffle[ToString /@ Attributes[Unevaluated[expr]], " "] //
-                StringJoin, HoldAll] @@ ToHeldExpression[str], "{" | "}"]
-
-isFunction[str_String] :=
-    With[{usg =
-        ToString[
-          Function[s, MessageName[s, "usage"], HoldAll] @@
-              ToHeldExpression[str]]},
-      str <> " " <>
-          If[StringMatchQ[usg, __ ~~ str ~~ "[" ~~ ___ ~~ "]" ~~ ___],
-            " = true", " = false"]
-    ]
-getAttributes[str_String] :=
-    StringTrim[
-      Function[expr,
-        Riffle[ToString /@ Attributes[Unevaluated[expr]], " "] //
-            StringJoin, HoldAll] @@ ToHeldExpression[str], "{" | "}"];
-getOptions[str_String] :=
-    StringTrim[
-      Function[expr,
-        Riffle[ToString[#, InputForm] & /@ (First /@
-            Options[Unevaluated[expr]]), " "] // StringJoin, HoldAll] @@
-          ToHeldExpression[str], "{" | "}" | ","]
+ClearAll[isFunction, getOptions, getAttributes];
+isFunction[symbol_String] := Not[TrueQ[Quiet@ToExpression[symbol, InputForm, ValueQ]]];
+getOptions[symbol_String /; isFunction[symbol]] := Quiet@Keys[ToExpression[symbol, InputForm, Options]];
+getOptions[__] := {};
+getAttributes[str_String /; isFunction[str]] := Quiet@ToExpression[str, InputForm, Attributes];
+getAttributes[__] := {};
 
 
 createInformation[name_String] := Module[{importance, info, context},