+++ src/de/halirutan/mathematica/codeinsight/completion/rendering/BuiltinSymbolLookupElement.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2017 Patrick Scheibe
+ * Copyright (c) 2018 Patrick Scheibe
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
@@ -7,16 +8,16 @@
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.halirutan.mathematica.codeinsight.completion.rendering;
@@ -33,7 +34,7 @@ import com.intellij.psi.PsiDocumentManager;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.ui.JBColor;
-import de.halirutan.mathematica.codeinsight.completion.SymbolInformationProvider.SymbolInformation;
+import de.halirutan.mathematica.information.impl.SymbolProperties;
 import de.halirutan.mathematica.lang.psi.api.FunctionCall;
 import de.halirutan.mathematica.settings.MathematicaSettings;
 import de.halirutan.mathematica.settings.MathematicaSettings.SmartEnterResult;
@@ -48,30 +49,31 @@ import org.jetbrains.annotations.Nullable;
 @SuppressWarnings("AnonymousClassVariableHidesContainingMethodVariable")
 public class BuiltinSymbolLookupElement extends LookupElement {
 
-  private final SymbolInformation myInfo;
+  private final SymbolProperties myInfo;
   private static final char OPEN_BRACKET = '[';
   private static final char CLOSING_BRACKET = ']';
 
-  public BuiltinSymbolLookupElement(SymbolInformation info) {
+  public BuiltinSymbolLookupElement(SymbolProperties info) {
     myInfo = info;
   }
 
   @NotNull
   @Override
   public String getLookupString() {
-    if ("System`".equals(myInfo.context)) {
-      return myInfo.nameWithoutContext;
+    if ("System`".equals(myInfo.getContext())) {
+      return myInfo.getName();
     }
-    return myInfo.name;
+    return myInfo.getContext() + myInfo.getName();
   }
 
   @Override
   public void renderElement(LookupElementPresentation presentation) {
+
     presentation.setItemText(getLookupString());
     presentation.setItemTextForeground(JBColor.blue);
-    presentation.setItemTextBold(myInfo.function);
-    presentation.setTailText(myInfo.function ? "[" + myInfo.getCallPattern() + "]" : "", true);
-    presentation.setTypeText(myInfo.context);
+    presentation.setItemTextBold(myInfo.isFunctionQ());
+    presentation.setTailText(myInfo.isFunctionQ() ? "[" + String.join(", ", myInfo.getCallPattern()) + "]" : "", true);
+    presentation.setTypeText(myInfo.getContext());
   }
 
   @Override
@@ -90,7 +92,7 @@ public class BuiltinSymbolLookupElement extends LookupElement {
     context.setAddCompletionChar(false);
 
     if (completionChar == Lookup.COMPLETE_STATEMENT_SELECT_CHAR) {
-      if (myInfo.function) {
+      if (myInfo.isFunctionQ()) {
         if (smartEnterSetting.equals(SmartEnterResult.INSERT_BRACES)) {
           document.insertString(context.getTailOffset(), "[]");
           final int currentPosition = context.getTailOffset();
@@ -98,9 +100,10 @@ public class BuiltinSymbolLookupElement extends LookupElement {
         } else if (smartEnterSetting.equals(SmartEnterResult.INSERT_CODE) || smartEnterSetting.equals(SmartEnterResult.INSERT_TEMPLATE)) {
           document.insertString(context.getTailOffset(), Character.toString(OPEN_BRACKET));
           final int currentPosition = context.getTailOffset();
-          document.insertString(currentPosition, myInfo.getCallPattern());
+          document.insertString(currentPosition, String.join(", ", myInfo.getCallPattern()));
           document.insertString(context.getTailOffset(), Character.toString(CLOSING_BRACKET));
-          final int endOffset = getFirstArgumentRange(myInfo) + currentPosition;
+          final int endOffset =
+              myInfo.getCallPattern().isEmpty() ? 0 : myInfo.getCallPattern().get(0).length() + currentPosition;
           editor.getSelectionModel().setSelection(currentPosition, endOffset);
           editor.getCaretModel().moveToOffset(endOffset);
 
@@ -156,12 +159,4 @@ public class BuiltinSymbolLookupElement extends LookupElement {
     PsiDocumentManager.getInstance(project).commitDocument(document);
   }
 
-  private int getFirstArgumentRange(SymbolInformation info) {
-    final String callPattern = info.getCallPattern();
-    final int firstComma = callPattern.indexOf(',');
-    if (firstComma == -1) {
-      return callPattern.length();
-    }
-    return firstComma;
-  }
 }