+++ src/de/halirutan/mathematica/information/impl/SymbolInformationImpl.java
@@ -22,10 +22,16 @@
 
 package de.halirutan.mathematica.information.impl;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import com.google.gson.stream.JsonReader;
 import com.intellij.openapi.diagnostic.Logger;
 import de.halirutan.mathematica.information.SymbolInformation;
 import org.jetbrains.annotations.NotNull;
 
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.regex.Pattern;
@@ -38,10 +44,24 @@ public class SymbolInformationImpl implements SymbolInformation {
   private static final Pattern namedCharacterPattern = Pattern.compile("\\\\\\[[A-Z][a-zA-Z]+]");
   private static Logger LOG = Logger.getInstance(SymbolInformation.class);
   private Map<String, String> myNamedCharacters = convertNamedCharacters();
-  private Set<String> myContexts =
-      convertResourceBundleToList("de.halirutan.mathematica.codeinsight.completion.MathematicaContexts");
-  private Set<String> myContextSymbols =
-      convertResourceBundleToList("de.halirutan.mathematica.codeinsight.completion.MathematicaContextSymbols");
+  private Set<String> myContexts;
+  private Set<String> myContextSymbols;
+  private Set<String> mySystemSymbols;
+  private Map<String, Double> mySymbolVersions;
+  private Map<String, SymbolProperties> mySymbolProperties;
+
+  public SymbolInformationImpl() {
+    try {
+      myContexts = loadJSONList("Contexts");
+      myContextSymbols =
+          loadJSONList("ContextSymbolNames");
+      mySystemSymbols = loadJSONList("SystemSymbolNames");
+      mySymbolProperties = loadSymbolProperties();
+      mySymbolVersions = loadSymbolVersions();
+    } catch (FileNotFoundException e) {
+      LOG.error("Cannot open Mathematica information resources", e);
+    }
+  }
 
 
   private static Map<String, String> convertNamedCharacters() {
@@ -56,10 +76,6 @@ public class SymbolInformationImpl implements SymbolInformation {
     return map;
   }
 
-  private static Set<String> convertResourceBundleToList(@NotNull String resource) {
-    ResourceBundle bundle = ResourceBundle.getBundle(resource);
-    return bundle.keySet();
-  }
 
   @Override
   public boolean isNamedCharacter(@NotNull String name) {
@@ -67,6 +83,46 @@ public class SymbolInformationImpl implements SymbolInformation {
       return myNamedCharacters.containsKey(name.substring(2, name.length() - 1));
     }
     return false;
+  }
+
+  @Override
+  public boolean isSystemSymbol(@NotNull String nameWithoutContext) {
+    return mySystemSymbols.contains(nameWithoutContext);
+  }
+
+  @Override
+  public boolean isBuiltinSymbol(@NotNull String nameWithContext) {
+    return myContextSymbols.contains(nameWithContext);
+  }
+
+  @Override
+  public Set<String> getContextSymbols() {
+    return myContextSymbols;
+  }
+
+  @Override
+  public Set<String> getAllContexts() {
+    return myContexts;
+  }
+
+  @Override
+  public Set<String> getSymbolsWithProperties() {
+    return mySymbolProperties.keySet();
+  }
+
+  @Override
+  public boolean hasProperties(@NotNull String nameWithContext) {
+    return mySymbolProperties.containsKey(nameWithContext);
+  }
+
+  @Override
+  public Collection<SymbolProperties> getAllSymbolProperties() {
+    return mySymbolProperties.values();
+  }
+
+  @Override
+  public SymbolProperties getSymbolProperties(@NotNull String nameWithContext) {
+    return mySymbolProperties.get(nameWithContext);
   }
 
   @Override
@@ -80,6 +136,52 @@ public class SymbolInformationImpl implements SymbolInformation {
   @Override
   public Set<String> getNamedCharacters() {
     return myNamedCharacters.keySet();
+  }
+
+  @Override
+  public double getSymbolVersion(@NotNull String nameWithContext) {
+    return mySymbolVersions.getOrDefault(nameWithContext, -1.0);
+  }
+
+  @Override
+  public boolean isVersionedSymbol(@NotNull String nameWithContext) {
+    return mySymbolVersions.containsKey(nameWithContext);
+  }
+
+  private HashSet<String> loadJSONList(String fileName) throws FileNotFoundException {
+    Gson gson = new Gson();
+    final URL resource = getClass().getClassLoader().getResource(
+        "de/halirutan/mathematica/codeinsight/completion/" + fileName + ".json");
+    if (resource == null) {
+      throw new FileNotFoundException(fileName);
+    }
+    JsonReader reader = new JsonReader(new FileReader(resource.getFile()));
+    return gson.fromJson(reader, new TypeToken<HashSet<String>>() {
+    }.getType());
+  }
+
+  private HashMap<String, Double> loadSymbolVersions() throws FileNotFoundException {
+    Gson gson = new Gson();
+    final URL resource = getClass().getClassLoader().getResource(
+        "de/halirutan/mathematica/codeinsight/completion/SymbolVersions.json");
+    if (resource == null) {
+      throw new FileNotFoundException("SymbolVersions.json");
+    }
+    JsonReader reader = new JsonReader(new FileReader(resource.getFile()));
+    return gson.fromJson(reader, new TypeToken<HashMap<String, Double>>() {
+    }.getType());
+  }
+
+  private HashMap<String, SymbolProperties> loadSymbolProperties() throws FileNotFoundException {
+    Gson gson = new Gson();
+    final URL resource = getClass().getClassLoader().getResource(
+        "de/halirutan/mathematica/codeinsight/completion/SymbolInformation.json");
+    if (resource == null) {
+      throw new FileNotFoundException("SymbolInformation.json");
+    }
+    JsonReader reader = new JsonReader(new FileReader(resource.getFile()));
+    return gson.fromJson(reader, new TypeToken<HashMap<String, SymbolProperties>>() {
+    }.getType());
   }
 
 }